First run at creating a program to calculate the number of weeks remaining until age 18 when given a childs birthday
Part of a project to emphasize the value of time when you are a parent.
Influencing your kids lives while they are at home is critical.
Eventually I want to create a visualization of this concept with
Marbles in a Jar as per the example in Losing Your Marbles by Reggie
Joiner.